Refactoring and change to global padding
